[svn-r20091] This is a fix for coverity bug #1683. Changed the two printfs to use %lu (unsigned long) for printing "dset_size".

+++ test/dsets.c
@@ -4099,7 +4099,7 @@ test_nbit_int_size(hid_t file)
    if((dset_size = H5Dget_storage_size(dataset)) < DSET_DIM1*DSET_DIM2*(precision/8) || 
        dset_size > DSET_DIM1*DSET_DIM2*(precision/8) + 1*KB) {
        H5_FAILED();
-       printf("    Line %d: wrong dataset size: %d\n",__LINE__, dset_size);
+       printf("    Line %d: wrong dataset size: %lu\n",__LINE__, (unsigned long)dset_size);
        goto error;
    } /* end if */
 
@@ -4306,7 +4306,7 @@ test_nbit_flt_size(hid_t file)
    if((dset_size = H5Dget_storage_size(dataset)) < DSET_DIM1*DSET_DIM2*(precision/8) || 
        dset_size > DSET_DIM1*DSET_DIM2*(precision/8) + 1*KB) {
        H5_FAILED();
-       printf("    Line %d: wrong dataset size: %d\n",__LINE__, dset_size);
+       printf("    Line %d: wrong dataset size: %lu\n",__LINE__, (unsigned long)dset_size);
        goto error;
    } /* end if */
